Gitty/requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
PyGithub==2.8.1
python-gitlab==8.0.0
